+++ hw-8/src/main/java/ru/dankoy/hw8/core/repository/commentary/CommentaryRepositoryCustom.java
@@ -5,6 +5,6 @@ import ru.dankoy.hw8.core.domain.Commentary;
 
 public interface CommentaryRepositoryCustom {
 
-  Optional<Commentary> getById(long id);
+  Optional<Commentary> getById(int id);
 
 }
