+++ hw-8/src/main/java/ru/dankoy/hw8/core/repository/genre/GenreRepositoryCustom.java
@@ -5,5 +5,5 @@ import ru.dankoy.hw8.core.domain.Genre;
 
 public interface GenreRepositoryCustom {
 
-  Optional<Genre> getById(long id);
+  Optional<Genre> getById(int id);
 }
